requirements.txt
asgiref==3.6.0
Celery==5.2.7
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-xml==2.0.0
djangorestframework-yaml==2.0.0
djangorestframework-simplejwt==5.2.2
django-extensions==3.2.1
django-debug-toolbar==3.8.1
django-admin-rangefilter==0.9.0
django-import-export
django-storages==1.13.2
django-filter==23.1
django-crispy-forms==2.0
crispy-bootstrap5==0.7
beautifulsoup4==4.12.0
sqlparse==0.4.3
ipython==8.10.0
requests
drf-yasg

#flake8
flake8
flake8-builtins
flake8-print
